Backend/SmartSolutions/requirements.txt
adagio==0.2.4
aiohttp==3.9.5
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.2.1
coreforecast==0.0.11
cycler==0.12.1
Cython==3.0.10
darts==0.26.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fs==2.4.16
fsspec==2024.6.1
fugue==0.9.1
holidays==0.53
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lightning-utilities==0.11.5
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
nfoursid==1.0.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==10.4.0
pmdarima==2.0.4
protobuf==5.27.2
pyarrow==17.0.0
pyod==2.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.3.2
scipy==1.14.0
shap==0.46.0
six==1.16.0
slicer==0.0.8
sqlparse==0.5.1
statsforecast==1.7.5
statsmodels==0.14.2
sympy==1.13.0
tbats==1.1.3
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==2.3.1
torchmetrics==1.4.0.post0
tqdm==4.66.4
triad==0.9.8
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
utilsforecast==0.1.12
xarray==2024.6.0
xgboost==2.1.0
yarl==1.9.4
